plumbing for packettracing
