+++ controller/pkg/packettracing/packettracing.go~
@@ -1,0 +1,10 @@
+package packettracing
+type tracingdirection int
+
+const (
+	Disabled              tracingdirection = 0
+	NetworkOnly           tracingdirection = 1
+	ApplicationOnly       tracingdirection = 2
+	NetworkAndApplication tracingdirection = 4
+	Invalid               tracingdirection = 8
+)
